safestringlib: unittests: cast pointer to intptr_t instead to int
Cast pointer intptr_t instead of int

warning: cast from pointer to integer of different size [-Wpointer-to-int-cast]

+++ unittests/test_strcasestr_s.c
@@ -131,7 +131,7 @@ int test_strcasestr_s()
 
     /* compare to legacy */
     std_sub = strcasestr(str1, str2);
-    if ((int)sub != (int)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
+    if ((intptr_t)sub != (intptr_t)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
         printf("%s %u  Error strcasestr_s() does not have same return as strcasestr() when str1 & str2 are zero length strings. rc=%u \n",
                      __FUNCTION__, __LINE__, rc);
         printf("str1:[%s]\n", str1);
@@ -158,7 +158,7 @@ int test_strcasestr_s()
 
     /* compare to legacy */
     std_sub = strcasestr(str1, str2);
-    if ((int)sub != (int)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
+    if ((intptr_t)sub != (intptr_t)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
         printf("%s %u  Error strcasestr_s() does not have same return value as strcasestr() when str2 is zero length string. rc=%u \n",
                              __FUNCTION__, __LINE__, rc);
         printf("str1:[%s]\n", str1);
@@ -186,7 +186,7 @@ int test_strcasestr_s()
 
     /* compare to legacy */
     std_sub = strcasestr(str1, str2);
-    if ((int)sub != (int)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
+    if ((intptr_t)sub != (intptr_t)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
         printf("%s %u  Error strcasestr_s() does not have same return value as strcasestr() when str2 is zero length string. rc=%u \n",
                                      __FUNCTION__, __LINE__, rc);
         printf("str1:[%s]\n", str1);
@@ -268,7 +268,7 @@ int test_strcasestr_s()
 
     /* compare to legacy */
     std_sub = strcasestr(str1, str2);
-    if ((int)sub != (int)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
+    if ((intptr_t)sub != (intptr_t)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
         printf("%s %u  Error strcasestr_s() does not have same return value as strcasestr() when str2 is substring of the end of str1. rc=%u \n",
                                      __FUNCTION__, __LINE__, rc);
         printf("str1:[%s]\n", str1);
@@ -299,7 +299,7 @@ int test_strcasestr_s()
 
     /* compare to legacy */
     std_sub = strcasestr(str1, str2);
-    if ((int)sub != (int)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
+    if ((intptr_t)sub != (intptr_t)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
         printf("%s %u  Error strcasestr_s() does not have same return value as strcasestr() when str2 is substring of middle of str1. rc=%u \n",
                                      __FUNCTION__, __LINE__, rc);
         printf("str1:[%s]\n", str1);
@@ -424,7 +424,7 @@ int test_strcasestr_s()
 
     /* compare to legacy */
     std_sub = strcasestr(str1, str2);
-    if ((int)sub != (int)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
+    if ((intptr_t)sub != (intptr_t)std_sub) { // comparison to handle 32-bit library return and 64-bit library return
         printf("%s %u  Error rc=%u \n",
                      __FUNCTION__, __LINE__, rc);
     }
